refactor: create Revision\Action namespace
Fix embeded actions broken, actionController is not only for exports
