+++ EMS/core-bundle/src/Resources/config/controllers.xml
@@ -237,15 +237,19 @@
         </service>
 
         <!-- Revision -->
-        <service id="EMS\CoreBundle\Controller\Revision\ActionController" public="true">
+        <service id="EMS\CoreBundle\Controller\Revision\Action\ActionController" public="true">
             <argument type="service" id="EMS\CoreBundle\Repository\TemplateRepository" />
             <argument type="service" id="EMS\CoreBundle\Service\EnvironmentService" />
             <argument type="service" id="ems.service.search" />
             <argument type="service" id="ems_common.pdf.printer.dom" />
             <argument type="service" id="ems_common.service.spreadsheet_generator_service" />
-            <argument type="service" id="ems_core.core_ui.ajax_service" />
             <argument type="service" id="logger" />
             <argument type="service" id="twig" />
+        </service>
+        <service id="EMS\CoreBundle\Controller\Revision\Action\ActionImportController" public="true">
+            <argument type="service" id="EMS\CoreBundle\Repository\TemplateRepository" />
+            <argument type="service" id="ems_core.core_ui.ajax_service" />
+            <argument type="service" id="form.factory" />
         </service>
 
         <service id="EMS\CoreBundle\Controller\Revision\DetailController" public="true">
