+++ EMS/core-bundle/src/Resources/config/routing/data.xml
@@ -59,12 +59,12 @@
            methods="GET|POST">
         <default key="public">0</default>
     </route>
-    <route id="emsco_data_action_modal_import" path="/action/modal-import/{environmentName}/{templateId}/{ouuid}"
-           controller="EMS\CoreBundle\Controller\Revision\ActionController::modalImport"
+    <route id="emsco_data_action_import" path="/action/import/{templateId}/{ouuid}"
+           controller="EMS\CoreBundle\Controller\Revision\Action\ActionImportController"
            methods="GET|POST"
         />
     <route id="emsco_data_private_action" path="/action/{environmentName}/{templateId}/{ouuid}/{_download}"
-           controller="EMS\CoreBundle\Controller\Revision\ActionController::render"
+           controller="EMS\CoreBundle\Controller\Revision\Action\ActionController::render"
            methods="GET">
         <default key="public">0</default>
         <default key="_download">0</default>
@@ -203,13 +203,13 @@
         <default key="public">0</default>
     </route>
     <route id="data.customview" path="/custom-view/{environmentName}/{templateId}/{ouuid}/{_download}"
-           controller="EMS\CoreBundle\Controller\Revision\ActionController::render"
+           controller="EMS\CoreBundle\Controller\Revision\Action\ActionController::render"
            methods="GET">
         <default key="public">0</default>
         <default key="_download">0</default>
     </route>
     <route id="ems_data_custom_template_protected" path="/template/{environmentName}/{templateId}/{ouuid}/{_download}"
-           controller="EMS\CoreBundle\Controller\Revision\ActionController::render"
+           controller="EMS\CoreBundle\Controller\Revision\Action\ActionController::render"
            methods="GET">
         <default key="public">0</default>
         <default key="_download">0</default>
