+++ EMS/core-bundle/src/Resources/views/action/modal_import.html.twig
@@ -2,7 +2,7 @@
 
 {% block body %}
     {% include '@EMSCore/core/alerts.twig' %}
-{#    {{ form(form) }}#}
+    {{ form(form) }}
 {% endblock %}
 
 {% block footer %}
